Function relocation
